Restructured the code to seperate meta sources
The meta sources are now in a seperate directory.

A new file is added to that directory `preconfig`. This file will handle
all the preconfigurations related to the metadata sources.

All the configs need to be defined inside the `CONFIG` class and the
values are later called in other files accordingly.

>NOTE: The `defaults` file is only for the default configurations of all
the supported values in the config. For all the default values of
metadata the `preconfig` file is to be used from now on.