Improvement for the playlist feature and bug fix
When a playlist is passed, the title is extracted automatically and even
though the title is tried to be sanitized as much as possible, it still
affects the metadata search.

So, a new feature is added. When the title is automatically extracted
then the user is asked if they want to change it or go with the extracted
one.

This improves the playlist extraction feature a lot.'

The bug fix is while setting the data, some providers do not have the
attribute provider so a check is added to see if they have that and
accordingly proceed.