Fix a bug of detecting if the passed URL is a valid playlist
The passed URL is checked to see if it is a valid playlist. Earlier it
was being done by basic `if present in <str>` which was causing a bug
of ignoring playlists that had just `http` and not `https`.

This is fixed by moving to using regex in order to make sure the passed
playlist is valid.