Disable progress when `--quiet` is passed.
When the `--quiet` flag is passed, disable the progress bar for the
download. This is also important because if `--quiet` is passed and the
output is being redirected to something other than stdout, then it will
end up raising an error (error directly from `downloader-cli`).

Disabling progress bar altogether is the expected behaviour and it stops
underlying errors as well.

Fixes #209